Rething game state
