+++ test/fungus_toast/games/grid_test.exs
@@ -1,6 +1,5 @@
 defmodule FungusToast.Games.GridTest do
   use ExUnit.Case, async: true
-  alias FungusToast.Games.Game.State.Grid
+  alias FungusToast.Games.Grid
   doctest Grid
-
 end
